CUDA: Start cleanup of backend
